+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # @biconomy/sdk
 
+## 0.0.9
+
+### Patch Changes
+
+- Added DAN helpers, keyGen + sigGen
+
 ## 0.0.8
 
 ### Patch Changes
