+++ src/sdk/clients/decorators/erc7579/getFallbackBySelector.ts
@@ -44,7 +44,7 @@ export async function getFallbackBySelector<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
