+++ src/sdk/clients/decorators/erc7579/getPreviousModule.ts
@@ -46,7 +46,7 @@ export async function getPreviousModule<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
