+++ src/sdk/clients/decorators/erc7579/installModule.ts
@@ -61,7 +61,7 @@ export async function installModule<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
