+++ src/sdk/clients/decorators/erc7579/installModules.ts
@@ -65,7 +65,7 @@ export async function installModules<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
