+++ src/sdk/clients/decorators/erc7579/isModuleInstalled.ts
@@ -57,7 +57,7 @@ export async function isModuleInstalled<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
