+++ src/sdk/clients/decorators/erc7579/supportsExecutionMode.ts
@@ -103,7 +103,7 @@ export async function supportsExecutionMode<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
