+++ src/sdk/clients/decorators/erc7579/supportsModule.ts
@@ -70,7 +70,7 @@ export async function supportsModule<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
