+++ src/sdk/clients/decorators/erc7579/uninstallFallback.ts
@@ -62,7 +62,7 @@ export async function uninstallFallback<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
