+++ src/sdk/clients/decorators/erc7579/uninstallModule.ts
@@ -64,7 +64,7 @@ export async function uninstallModule<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
