+++ src/sdk/clients/decorators/erc7579/uninstallModules.ts
@@ -61,7 +61,7 @@ export async function uninstallModules<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
