+++ src/sdk/clients/decorators/index.ts
@@ -1,3 +1,4 @@
 export * from "./erc7579"
 export * from "./smartAccount"
 export * from "./bundler"
+export * from "./dan"
