+++ src/sdk/clients/decorators/smartAccount/sendTransaction.ts
@@ -60,7 +60,7 @@ export async function sendTransaction<
 
     if (!account_) {
       throw new AccountNotFoundError({
-        docsPath: "/docs/actions/wallet/sendTransaction"
+        docsPath: "/nexus-client/methods#sendtransaction"
       })
     }
