+++ src/sdk/modules/ownableValidator/decorators/addOwner.ts
@@ -54,7 +54,7 @@ export async function addOwner<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
