+++ src/sdk/modules/ownableValidator/decorators/getAddOwnerTx.ts
@@ -45,7 +45,7 @@ export async function getAddOwnerTx<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
