+++ src/sdk/modules/ownableValidator/decorators/getOwners.ts
@@ -39,7 +39,7 @@ export async function getOwners<
   // Review docspath below.
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
