+++ src/sdk/modules/ownableValidator/decorators/getRemoveOwnerTx.ts
@@ -41,7 +41,7 @@ export async function getRemoveOwnerTx<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
