+++ src/sdk/modules/ownableValidator/decorators/getSetThresholdTx.ts
@@ -50,7 +50,7 @@ export async function getSetThresholdTx<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
