+++ src/sdk/modules/ownableValidator/decorators/getThreshold.ts
@@ -56,7 +56,7 @@ export async function getThreshold<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
