+++ src/sdk/modules/ownableValidator/decorators/removeOwner.ts
@@ -65,7 +65,7 @@ export async function removeOwner<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
