+++ src/sdk/modules/ownableValidator/decorators/setThreshold.ts
@@ -69,7 +69,7 @@ export async function setThreshold<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
