+++ src/sdk/modules/smartSessionsValidator/decorators/trustAttesters.ts
@@ -75,7 +75,7 @@ export async function trustAttesters<
 
   if (!account_) {
     throw new AccountNotFoundError({
-      docsPath: "/nexus/nexus-client/methods#sendtransaction"
+      docsPath: "/nexus-client/methods#sendtransaction"
     })
   }
